Split Match struct into its own file
